+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Fix `add_reference_concurrently` to check for mismatched key types
+
 ## 0.14.1 (2024-02-21)
 
 - Fix `MigrationRunner` to consider `run_background_migrations_inline` proc